20139 - move step_function testcases
